typescript integration WIP
