+++ .gitignore
@@ -1,4 +1,6 @@
 .idea
 src/credentials.js
+dest/
+typings/
 jspm_packages
 node_modules
