+++ index.html
@@ -15,7 +15,7 @@
     <script src="./jspm_packages/system.js"></script>
     <script src="./config.js"></script>
     <script>
-      System.import('src/bootstrap.js');
+      System.import('dest/bootstrap.js');
     </script>
 
   </head>
