+++ src/components/footer/footer.tsx
@@ -1,10 +1,10 @@
 'use strict';
 
-import React from 'react';
+import * as React from 'react';
 
 import './footer.css!';
 
-class Footer extends React.Component {
+class Footer extends React.Component <MyProps, MyState>  {
 
   render() {
     return (
