+++ src/components/header/header.tsx
@@ -1,6 +1,6 @@
 'use strict';
 
-import React from 'react';
+import * as React from 'react';
 
 import './header.css!';
