+++ src/components/navigation/navigation.tsx
@@ -1,6 +1,6 @@
 'use strict';
 
-import React from 'react';
+import * as React from 'react';
 import { Link } from 'react-router';
 
 import './navigation.css!';
