+++ src/components/user-details/user-details.tsx
@@ -1,6 +1,6 @@
 'use strict';
 
-import React from 'react';
+import * as React from 'react';
 
 import './user-details.css!';
 import { GithubStore } from '../../stores/github/github-store';
