+++ typings.json
@@ -1,0 +1,10 @@
+{
+  "name": "github-dashboard",
+  "version": false,
+  "dependencies": {},
+  "ambientDependencies": {
+    "react": "registry:dt/react#0.14.0+20160423065914",
+    "react-dom": "registry:dt/react-dom#0.14.0+20160412154040",
+    "react-router": "registry:dt/react-router#2.0.0+20160423062133"
+  }
+}
